+++ extensions/tiktok/src/main/java/app/morphe/extension/tiktok/settings/SettingsStatus.java
@@ -29,6 +29,7 @@ public class SettingsStatus {
     public static boolean blockAuthorEnabled = false;
     public static boolean inboxFilterEnabled = false;
     public static boolean commentToolsEnabled = false;
+    public static boolean videoOverlaysEnabled = false;
 
     public static void enableFeedFilter() {
         feedFilterEnabled = true;
@@ -120,6 +121,10 @@ public class SettingsStatus {
 
     public static void enableCommentTools() {
         commentToolsEnabled = true;
+    }
+
+    public static void enableVideoOverlays() {
+        videoOverlaysEnabled = true;
     }
 
     public static void load() {